requirements.txt
streamlit>=1.31.0
cohere>=5.5.7
python-dotenv>=1.0.1
PyMuPDF>=1.24.7
numpy>=1.26.0
